+++ CHANGES.md
@@ -1,4 +1,7 @@
 ### Updates from version 1.5.6 to 1.5.7
+* added support for Windsat Coriolis data
+
+### Updates from version 1.5.6 to 1.5.7
 * added support for SMOS L1C daily aggregated products
 * added support for SIC-CCI RRDP insitu data
 * Corrected reading bug in slstr-s3x-uor reader